+++ restler-service/src/main/java/net/researchgate/restler/service/mongo/MongoModule.java
@@ -3,7 +3,7 @@ package net.researchgate.restler.service.mongo;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
-import com.mongodb.MongoClient;
+import com.mongodb.client.MongoClient;
 import dev.morphia.Datastore;
 import dev.morphia.Morphia;
 import org.slf4j.Logger;
@@ -14,13 +14,12 @@ public class MongoModule extends AbstractModule {
     private static final Logger LOGGER = LoggerFactory.getLogger(MongoModule.class);
 
     private final MongoClient mongoClient;
-    private final Morphia morphia;
-    private final String databaseName;
+
+    private final Datastore datastore;
 
     public MongoModule(MongoClient mongoClient, String databaseName) {
         this.mongoClient = mongoClient;
-        this.morphia = new Morphia();
-        this.databaseName = databaseName;
+        this.datastore = Morphia.createDatastore(mongoClient, databaseName);
     }
 
     @Override
@@ -33,15 +32,8 @@ public class MongoModule extends AbstractModule {
     }
 
     @Provides
-    Morphia provideMorphia() {
-        LOGGER.info("Providing morphia object");
-        return morphia;
-    }
-
-    @Provides
     @Singleton
     Datastore provideDataStore() {
-        LOGGER.info("Creating new morphia datastore: {}", databaseName);
-        return morphia.createDatastore(mongoClient, databaseName);
+        return datastore;
     }
 }